Change data type of space to 64 bit number

+++ db/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE Machines (
 
 CREATE TABLE Disks (
   id serial PRIMARY KEY,
-  space integer NOT NULL,
+  space bigint NOT NULL,
   machineId integer NULL
 );
